Bug 1142956 - [Shared] Consolidate duplicated async_semaphore.js file
