+++ apps/system/index.html
@@ -14,6 +14,7 @@
     <script src="shared/js/l10n_date.js"></script>
     <script src="shared/js/idletimer.js"></script>
     <script src="shared/js/dump.js"></script>
+    <script defer src="shared/js/async_semaphore.js"></script>
     <script defer src="shared/js/nfc_utils.js"></script>
     <script defer src="shared/js/event_safety.js"></script>
     <script defer src="shared/js/gesture_detector.js"></script>
@@ -141,7 +142,6 @@
     <script defer src="js/browser_settings.js"></script>
     <script defer src="js/bootstrap.js"></script>
     <script defer src="js/screenshot.js"></script>
-    <script defer src="js/async_semaphore.js"></script>
     <script defer src="js/sound_manager.js"></script>
     <script defer src="js/source_view.js"></script>
     <script defer src="js/browser_key_event_manager.js"></script>
