+++ apps/system/js/sound_manager.js
@@ -273,7 +273,7 @@
       this.muteIcon = new MuteIcon(this);
       this.muteIcon.start();
     }.bind(this)).catch(function(err) {
-      console.error(err); 
+      console.error(err);
     });
 
     // mozChromeEvent fired from Gecko is earlier been loaded,
