+++ tv_apps/app-deck/index.html
@@ -24,6 +24,7 @@
     <script src="shared/js/l10n.js"></script>
     <script src="shared/js/manifest_helper.js"></script>
     <script defer src="bower_components/evt/index.js"></script>
+    <script defer src="shared/js/async_semaphore.js"></script>
     <script defer src="shared/js/smart-screen/scrollable.js"></script>
     <script defer src="shared/js/smart-screen/spatial_navigator.js"></script>
     <script defer src="shared/js/smart-screen/key_navigation_adapter.js"></script>
@@ -35,7 +36,6 @@
     <script defer src="shared/js/smart-screen/cards/application.js"></script>
     <script defer src="shared/js/smart-screen/cards/deck.js"></script>
     <script defer src="shared/js/smart-screen/cards/folder.js"></script>
-    <script defer src="shared/js/smart-screen/async_semaphore.js"></script>
     <script defer src="shared/js/smart-screen/card_manager.js"></script>
 
     <!-- building blocks -->
