+++ tv_apps/smart-home/_test/unit/card_manager_test.js
@@ -3,6 +3,7 @@
           MockCardStore, MockPipedPromise, MockXMLHttpRequest */
 
 require('/bower_components/evt/index.js');
+require('/shared/js/async_semaphore.js');
 require('/shared/js/uuid.js');
 require('/shared/js/smart-screen/shared_utils.js');
 require('/shared/test/unit/mocks/mock_piped_promise.js');
@@ -12,7 +13,6 @@ require('/shared/js/smart-screen/cards/card.js');
 require('/shared/js/smart-screen/cards/deck.js');
 require('/shared/js/smart-screen/cards/folder.js');
 require('/shared/js/smart-screen/cards/application.js');
-require('/shared/js/smart-screen/async_semaphore.js');
 
 suite('smart-home/CardManager', function() {
   var realPipedPromise;
