+++ tv_apps/smart-home/index.html
@@ -28,6 +28,7 @@
     <script defer src="shared/js/smart-screen/init_gesture.js"></script>
 
     <!-- Shared libraries -->
+    <script defer src="shared/js/async_semaphore.js"></script>
     <script defer src="shared/js/uuid.js"></script>
     <script defer src="shared/js/iac_handler.js"></script>
 
@@ -41,7 +42,6 @@
     <script defer src="shared/js/smart-screen/cards/application.js"></script>
     <script defer src="shared/js/smart-screen/cards/deck.js"></script>
     <script defer src="shared/js/smart-screen/cards/folder.js"></script>
-    <script defer src="shared/js/smart-screen/async_semaphore.js"></script>
     <script defer src="shared/js/smart-screen/card_manager.js"></script>
     <script defer src="shared/js/smart-screen/animations.js"></script>
