+++ tv_apps/smart-system/index.html
@@ -126,7 +126,7 @@
     <script defer src="js/external_storage_monitor.js"></script>
     <script defer src="js/bootstrap.js"></script>
     <script defer src="js/screenshot.js"></script>
-    <script defer src="shared/js/smart-screen/async_semaphore.js"></script>
+    <script defer src="shared/js/async_semaphore.js"></script>
     <script defer src="js/sound_manager.js"></script>
     <script defer src="js/browser_key_event_manager.js"></script>
     <script defer src="js/hardware_buttons.js"></script>
